laba9: complete

+++ labs/src/laba8/UserList.jsx
@@ -1,10 +1,12 @@
 import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
 import { getUsers, deleteUser, blockUser } from 'D:/React/ssrvp/labs/src/laba6/api.js';
 import UserTable from 'D:/React/ssrvp/labs/src/laba8/UserTable.jsx';
 import { Box, Typography } from '@mui/material';
 
 export default function UserList() {
   const [users, setUsers] = useState([]);
+  const currentUserId = useSelector((state) => state.auth.currentUserId);
 
   useEffect(() => {
     refreshUsers();
@@ -16,11 +18,19 @@ export default function UserList() {
   };
 
   const handleDelete = async (id) => {
+    if (id === currentUserId) {
+      alert("Вы не можете удалить сами себя!");
+      return;
+    }
     await deleteUser(id);
     refreshUsers();
   };
 
   const handleBlock = async (id, blocked) => {
+    if (id === currentUserId) {
+      alert("Вы не можете заблокировать сами себя!");
+      return;
+    }
     await blockUser(id, blocked);
     refreshUsers();
   };
